Create Hospital Compare Data Analysis
This code introduces a few functions that can be useful to analyze the Data from the outcome-of-care-measures.csv file, provided by Hospital Compare. Namely, the functions aim to help find hospitals with a given ranking in terms of how low their mortality rate for a given illness (outcome) is, filtering by state.